requirements.txt
python-telegram-bot==13.15
apscheduler
matplotlib

requests
